requirements.txt
tensorflow==2.1.0
tensorflow-model-optimization==0.5.0
matplotlib
scikit-learn
pandas
progressbar2
seaborn>=0.11.1
opencv-python>=4.5